updated main; split unittest for main and environment
