+++ ecosystem/__init__.py
@@ -1,1 +1,1 @@
-from main import ecosystem
+from main import ecosystem, list_available_tools
