+++ ecosystem/main.py
@@ -78,7 +78,7 @@ def ecosystem(tools=None, run_application=None, set_environment=False, force_reb
 
             call_process(MAKE_COMMAND)
 
-    elif run_application:
+    elif run_application is not None:
         env = Environment(tools)
         if env.success:
             env.set_env(os.environ)
@@ -131,19 +131,22 @@ Example:
         return 0
 
     tools = args.tools.split(',') if args.tools is not None else []
-    run_application = args.run
-    set_environment = args.setenv
-    force_rebuild = args.force
-    quick_build = args.make
-    run_build = args.build
-    deploy = args.deploy
-    if deploy:
-        force_rebuild = True
-        run_build = True
-        quick_build = False
+    # run_application = args.run
+    # set_environment = args.setenv
+    # force_rebuild = args.force
+    # quick_build = args.make
+    # run_build = args.build
+    # deploy = args.deploy
+    # if deploy:
+    #     force_rebuild = True
+    #     run_build = True
+    #     quick_build = False
 
     try:
-        ecosystem(tools, run_application, set_environment, force_rebuild, quick_build, run_build, deploy)
+        if args.deploy:
+            ecosystem(tools, args.run, args.setenv, True, False, True, args.deploy)
+        else:
+            ecosystem(tools, args.run, args.setenv, args.force, args.make, args.build, args.deploy)
         return 0
     except Exception, e:
         sys.stderr.write('ERROR: {0:s}'.format(str(e)))
