+++ tests/test_environment.py
@@ -1,12 +1,11 @@
 import unittest
 import os
-from ecosystem.main import list_available_tools
 from ecosystem.environment import ValueWrapper, Variable, Tool, Environment
 
 ECO_ROOT = os.environ.get('ECO_ROOT') or os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 
 
-class ValueWrapperTester(unittest.TestCase):
+class TestValueWrapper(unittest.TestCase):
 
     # def setUp(self):
     #     self.value_wrapper_obj = ValueWrapper()
@@ -57,7 +56,7 @@ class ValueWrapperTester(unittest.TestCase):
         self.assertEqual(value_wrapper_obj.absolute_value, False)
 
 
-class VariableTester(unittest.TestCase):
+class TestVariable(unittest.TestCase):
 
     def _test_append_value(self, variable, value,
                            dependents=None,
@@ -125,7 +124,7 @@ class VariableTester(unittest.TestCase):
         self.assertEqual(self.variable_obj.env, '/some/path')
 
 
-class ToolTester(unittest.TestCase):
+class TestTool(unittest.TestCase):
 
     def setUp(self):
         self.environ = os.environ.copy()
@@ -170,7 +169,7 @@ class ToolTester(unittest.TestCase):
     #     self.assertTrue(self.tool_obj.definesVariable('foo'), False)
 
 
-class EnvironmentTester(unittest.TestCase):
+class TestEnvironment(unittest.TestCase):
 
     def setUp(self):
         self.environ = os.environ.copy()
@@ -236,33 +235,6 @@ setenv PATH ${MAYA_LOCATION}/bin:${YETI_ROOT}/bin:${PATH}
         # elif current_os in ['linux', 'linux2']:
         #     test_get_env = None
         self.assertEqual(self.environment_obj.get_env(), test_get_env)
-
-
-class ListAvailableToolsTester(unittest.TestCase):
-
-    def setUp(self):
-        self.environ = os.environ.copy()
-        os.environ['ECO_ENV'] = os.path.join(ECO_ROOT, 'tests', 'test_env')
-        self.last_tool = 'yeti1.3.16'
-        self.no_tools = 2
-
-    def tearDown(self):
-        os.environ = self.environ
-
-    def test_list_available_tools(self):
-        available_tools = list_available_tools()
-        self.assertEqual(available_tools[-1], self.last_tool)
-        self.assertEqual(len(available_tools), self.no_tools)
-
-
-# class MainTester(unittest.TestCase):
-#
-#     def setUp(self):
-#         self.environ = os.environ.copy()
-#         os.environ['ECO_ENV'] = os.path.join(ECO_ROOT, 'env')
-#
-#     def tearDown(self):
-#         os.environ = self.environ
 
 
 if __name__ == '__main__':
